Update on libk
Added %x function to printf, added clean to iso.sh updated makefile
warnings and flags

+++ system/include/kernel/drivers/vga.h
@@ -29,7 +29,7 @@ namespace driver_vga {
 		};
 
 		static inline uint8_t vga_entry_color(enum vga_color fg, enum vga_color bg) {
-			return fg | bg << 4;
+			return (uint8_t) fg | bg << 4;
 		}
 
 		static inline uint16_t vga_entry(unsigned char uc, uint8_t color) {
